Allow print() to be used for debugging analysis server.
We now run analysis server inside a zone so that all print() calls it
executes can be captured to a buffer (rather than being sent to
standard output, where they will confuse the client).  When the server
is run in HTTP server mode, the print buffer is displayed on the
status page.

R=brianwilkerson@google.com, danrubel@google.com

Review URL: https://codereview.chromium.org//686763004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@41380 260f80e4-7a28-3924-810f-c04153c831b5