Always check if instance method name collides with superclass static.
Previously we only did this check if nothing was overridden.
Persumably that was intended as a speed optimization, since at first
glance it would seem that if the instance method overrides something,
then it can't possibly conflict with a static method in the
superclass.  However, that optimization doesn't work if the method
being overridden is in an interface (in that case the superclass can
still have a static method with the same name).  So just go ahead and
check in all cases.

BUG=dartbug.com/18947
R=scheglov@google.com

Review URL: https://codereview.chromium.org//698973005

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@41501 260f80e4-7a28-3924-810f-c04153c831b5