Don't to do HTML entity decoding for multi-part form fields
The code was trying to do HTML entity decoding on text field values
in multipart form posts. This has been removed as it did not work.

The parsing was expecting HTML entities to always be used, but that is
not the case. Chrome, Firefox and Safari will do it if e.g. the attribute
accept-charset="latin1" is set on the HTML form element, but the default
when accept-charset is not set or is set to utf8 UTF-8 encoding will be
used.

Also when HTML entity encoding is used, the encoding done by the browser
cannot be consistently  decoded. E.g. if the value '&#<some large char>' is
entered in a form with accept-charset="latin1" the value in the
multipart/form-data response is '&#&#12402;' (<some large char> is HTML
entity encoded as &#12402;), so '&#' is not escaped in any way.

Passing the value returned by the browser into user code without trying to
be clever is the right thing to do.

R=kustermann@google.com
BUG=

Review URL: https://codereview.chromium.org//730203008

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@41821 260f80e4-7a28-3924-810f-c04153c831b5