Remove flaky check from reanalyze_concurrent_test.
This should avoid sporadic false failures with this test.  If in the
future we genuinely regress bug 21448, there is a danger that this
test won't catch it reliably, but that is mitigated by the fact that
(a) even flaky failures do get noticed, and (b) the bug was severe
enough that it probably would be noticed in manual testing.  The extra
risk seems worth it to avoid the sporadic false failures we've been
seeing.

BUG=dartbug.com/21638
R=scheglov@google.com

Review URL: https://codereview.chromium.org//724403003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@41870 260f80e4-7a28-3924-810f-c04153c831b5