Annotate remaining analyzer/analysis server tests with "@ReflectiveTestCase()"
In r42181, I failed to notice the need to annotate these tests because
they aren't currently being run under dart2js (most likely due to
slowness).  I've modified runReflectiveTests() to check for the
presence of the annotation before running the tests, in order to
ensure that all the classes are properly annotated now, and so that we
won't forget to add annotations in the future.

R=scheglov@google.com

Review URL: https://codereview.chromium.org//742273005

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@42185 260f80e4-7a28-3924-810f-c04153c831b5