dart2js: accept const map literals as expression statements.
Dart2js was broken for:

    <String, int> {};

The reason was that we tried to find an identifier after an initial type (looking for something like `Set<int> foo`) to parse a declaration. We assumed that the first token would be a type name and skipped the `<` when we did this search, thus believing that there was a declaration.

BUG= http://dartbug.com/12793
BUG= http://dartbug.com/12794
BUG= http://dartbug.com/21155
R=ahe@google.com

Review URL: https://codereview.chromium.org//796363003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@42356 260f80e4-7a28-3924-810f-c04153c831b5