dart2js: Change typedef encoding.
Currently typedefs have a descriptor field that looks like they don't have any super-class:

_Equality: {
  "^": ":58"
}

When we call finishClass we thus assume that it's the Dart Object class. While this didn't hurt (as far as I could tell), it feels wrong.

Since typedefs are only emitted when reflection is enabled I'm ok with adding more code to the output (not that it's that much).

The output now looks as follows:

  _Equality: {
    "^": "Object;",
    $typedefType: 58,
    $$isTypedef: true
  },

At the same time I removed the shared encoding class. It turns out that it doesn't give a lot of flexibility, and by using the shared embedded constant strings, the linkage is not too bad.

R=karlklose@google.com

Review URL: https://codereview.chromium.org//804423003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@42437 260f80e4-7a28-3924-810f-c04153c831b5