Report the correct number of tests when tests are skipped due to compile time errors.
The current setup will count tests that are skipped when runtime is a browser due to compile time errors twice. Example:
tools/test.py -mrelease -rdrt -cdart2js --report language/bad_override
<snip>
Total: 10 tests
 * 0 tests will be skipped (0 skipped by design)
 * 0 tests are expected to be flaky but not crash
 * 5 tests are expected to pass
 * 0 tests are expected to fail that we won't fix
 * 2 tests are expected to fail that we should fix
 * 0 tests are expected to crash that we should fix
 * 0 tests are allowed to timeout
 * 3 tests are skipped on browsers due to compile-time error

This is a multitest that has 7 tests (including none). 3 of these are giving compile-time errors, 2 of these are failing (as a compile-time error when they should not, but that is not the point), 2 are passing.

This is preparation of a bigger refactoring of the summary reporting, so that I can have sane numbers for validation.

R=kustermann@google.com

Review URL: https://codereview.chromium.org//816703003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@42460 260f80e4-7a28-3924-810f-c04153c831b5