Allow LetCont to bind multiple continuations.
LetCont can now bind multiple continuations in the same scope.  As an
example, the 'then' and 'else' continuations of a branch are siblings
in the expression, and neither is in the scope of the other.  None of
the LetCont-bound continuations are in scope for any of the
continuation bodies.  Recursive continuations are still recursive
values, not recursive bindings, so there is no direct way to have
mutually recursive continuations.  This matches the JavaScript
semantics.

R=asgerf@google.com, sigurdm@google.com, karlklose@google.com
BUG=

Review URL: https://codereview.chromium.org//848363002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@42945 260f80e4-7a28-3924-810f-c04153c831b5