Handle super-method invocations in CPS->JS backend.
InvokeSuperMethod has been replaced by InvokeMethodDirectly. This
instruction works like InvokeMethod, except it has an explicit target
element that it always invokes, regardless of overriding.

The new instruction is easier to handle in the JS backend, it is more
amenable to inlining, and it should also be useful for calling
backend-specific methods, such as constructor bodies.

When targeting Dart, we translate it back to a 'super' at the end,
because the instruction is never used for any other purpose with
that backend.

R=karlklose@google.com, sigurdm@google.com

Review URL: https://codereview.chromium.org//861713002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@43011 260f80e4-7a28-3924-810f-c04153c831b5